requirements.txt
ccxt
pandas
numpy
anthropic
ta
rich
flask
apscheduler
markdown
yfinance
pytz
openai
schedule